requirements.txt
# Core dependencies
streamlit>=1.24.0
langchain>=0.0.300
langchain-openai>=0.0.2
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Database connectors
pymongo>=4.5.0
mysql-connector-python>=8.2.0
snowflake-connector-python>=3.5.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# File handling
pandas>=2.1.0
openpyxl>=3.1.2
xlrd>=2.0.1

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0
pre-commit>=3.4.0

# Type stubs for better type checking
types-requests>=2.31.0.2
types-python-dateutil>=2.8.19.14
types-PyYAML>=6.0.12.12